% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse.R
\name{collapse4_pop}
\alias{collapse4_pop}
\title{Collapse Population Race Categories into 4 Categories}
\usage{
collapse4_pop(.data, prefix = "pop_")
}
\arguments{
\item{.data}{tibble, data.frame, or sf tibble}

\item{prefix}{Default is \code{pop_}. The prefix for the race categories.}
}
\value{
.data with columns collapsed
}
\description{
Collapses Other, AIAN, Asian, NHPI, and Two+ into other.
}
\examples{
data(mt_county)
mt_county <- mt_county \%>\% collapse4_pop()
}
\concept{collapse}
