% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse.R
\name{collapse5}
\alias{collapse5}
\title{Collapse Full Race Categories into 5 Categories}
\usage{
collapse5(.data, prefix)
}
\arguments{
\item{.data}{tibble, data.frame, or sf tibble}

\item{prefix}{The prefix(es) for the race categories. Must be a character vector.}
}
\value{
.data with columns collapsed
}
\description{
Collapses Other, AIAN, NHPI, and Two+ into Other, by prefix.
}
\examples{
data(mt_county)
mt_county <- mt_county \%>\% collapse5(prefix = c('pop_', 'vap_'))
}
\concept{collapse}
