% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoid.R
\name{custom_geoid}
\alias{custom_geoid}
\title{Create a GEOID from Columns}
\usage{
custom_geoid(.data, ...)
}
\arguments{
\item{.data}{dataframe, tibble, or sf tibble}

\item{...}{columns of .data in the order you want to make the GEOID}
}
\value{
.data with new column GEOID
}
\description{
Create a GEOID from Columns
}
\examples{
data(mt_county)
mt_county <- mt_county \%>\% custom_geoid(GEOID)
}
\concept{geoid}
