% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{fips_2010}
\alias{fips_2010}
\title{Counties FIPS 2010}
\usage{
data('fips_2010')
}
\value{
tibble
}
\description{
Contains three columns:
\itemize{
\item state: state FIPS
\item county: county FIPS
\item name: county name
}
}
\examples{
data('fips_2010')

}
\concept{data}
