% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{fips_2020}
\alias{fips_2020}
\title{Counties FIPS 2020}
\usage{
data('fips_2020')
}
\value{
tibble
}
\description{
Contains three columns:
\itemize{
\item state: state FIPS
\item county: county FIPS
\item name: county name
}
}
\examples{
data('fips_2020')
}
\concept{data}
