% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join.R
\name{join_abb_ansi}
\alias{join_abb_ansi}
\title{Join Abb by ANSI}
\usage{
join_abb_ansi(.data, .ansi)
}
\arguments{
\item{.data}{data.frame or tibble}

\item{.ansi}{column with state ansi}
}
\value{
.data with column .ansi replaced with state abbreviation
}
\description{
Adds a column with state abbreviation joining by a column with state ansi
}
\examples{
data('stata')
stata \%>\% join_abb_ansi(ansi)
}
\concept{join}
