% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join.R
\name{join_abb_fips}
\alias{join_abb_fips}
\title{Join Abb by FIPS}
\usage{
join_abb_fips(.data, .fips)
}
\arguments{
\item{.data}{data.frame or tibble}

\item{.fips}{column with state fips}
}
\value{
.data with column .fips replaced with state abb
}
\description{
Adds a column with state abbreviation joining by a column with state fips
}
\examples{
data('stata')
stata \%>\% join_abb_fips(fips)
}
\concept{join}
