% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join.R
\name{join_ansi_abb}
\alias{join_ansi_abb}
\title{Join ANSI by Abb}
\usage{
join_ansi_abb(.data, .abb)
}
\arguments{
\item{.data}{data.frame or tibble}

\item{.abb}{column with state abbrevaition}
}
\value{
.data with column .abb replaced with state ansi
}
\description{
Replaces state abbreviation with state ansi
}
\examples{
data('stata')
stata \%>\% join_ansi_abb(abb)
}
\concept{join}
