% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join.R
\name{join_ansi_fips}
\alias{join_ansi_fips}
\title{Join ANSI by FIPS}
\usage{
join_ansi_fips(.data, .fips)
}
\arguments{
\item{.data}{data.frame or tibble}

\item{.fips}{column with state fips}
}
\value{
.data with column .fips replaced with state ansi
}
\description{
Adds a column with state ansi joining by a column with state fips
}
\examples{
data('stata')
stata \%>\% join_ansi_fips(fips)
}
\concept{join}
