% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join.R
\name{join_ansi_name}
\alias{join_ansi_name}
\title{Join ANSI by Name}
\usage{
join_ansi_name(.data, .name)
}
\arguments{
\item{.data}{data.frame or tibble}

\item{.name}{column with state ansi}
}
\value{
.data with column .name replaced with ansi
}
\description{
Replaces state name with state ansi
}
\examples{
data('stata')
stata \%>\% join_ansi_name(name)
}
\concept{join}
