% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join.R
\name{join_fips_ansi}
\alias{join_fips_ansi}
\title{Join FIPS by ANSI}
\usage{
join_fips_ansi(.data, .ansi)
}
\arguments{
\item{.data}{data.frame or tibble}

\item{.ansi}{column with state ansi}
}
\value{
.data with column .ansi replaced with state fips
}
\description{
Adds a column with state fips joining by a column with state ansi
}
\examples{
data('stata')
stata \%>\% join_fips_ansi(ansi)
}
\concept{join}
