% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join.R
\name{join_fips_name}
\alias{join_fips_name}
\title{Join FIPS by Name}
\usage{
join_fips_name(.data, .name)
}
\arguments{
\item{.data}{data.frame or tibble}

\item{.name}{column with state name}
}
\value{
.data with column .name replaced with fips
}
\description{
Adds a column with state fips joining by a column with state name
}
\examples{
data('stata')
stata \%>\% join_fips_name(name)
}
\concept{join}
