% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join.R
\name{join_name_abb}
\alias{join_name_abb}
\title{Join Name to Abb}
\usage{
join_name_abb(.data, .name)
}
\arguments{
\item{.data}{data.frame or tibble}

\item{.name}{column with state abbreviation}
}
\value{
.data with column .name replaced with abbreviation
}
\description{
Adds a column with state abbs joining to a column with state names
}
\examples{
data('stata')
stata \%>\% join_name_abb(name)
}
