% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join.R
\name{join_name_ansi}
\alias{join_name_ansi}
\title{Join Name by ANSI}
\usage{
join_name_ansi(.data, .ansi)
}
\arguments{
\item{.data}{data.frame or tibble}

\item{.ansi}{column with state ansi}
}
\value{
.data with column .ansi replaced with state name
}
\description{
Adds a column with state name joining by a column with state ansi
}
\examples{
data('stata')
stata \%>\% join_name_ansi(name)
}
\concept{join}
