% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_key.R
\name{key}
\alias{key}
\alias{has_census_key}
\alias{get_census_key}
\title{Check or Get Census API Key}
\usage{
has_census_key()

get_census_key(key = "")
}
\arguments{
\item{key}{Census API Key as a character}
}
\value{
logical if \code{has}, key if \code{get}
}
\description{
Check or Get Census API Key
}
\examples{
has_census_key()
}
