% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{loop_dec_counties}
\alias{loop_dec_counties}
\title{Loop over Counties Decennial}
\usage{
loop_dec_counties(.geography, .state, .year, .geometry, .vars)
}
\description{
Loop over Counties Decennial
}
\keyword{internal}
