% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.R
\name{match_ansi}
\alias{match_ansi}
\title{Try to Match to State ANSI}
\usage{
match_ansi(state)
}
\arguments{
\item{state}{character with state FIPS, Abbreviation, Name, or ANSI}
}
\value{
ANSI if a match is found or character(0) if no match is found
}
\description{
Searches for an exact match and offers the best match if no exact match
}
\examples{
match_ansi('NY')
match_ansi('01')
}
\concept{match}
