% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.R
\name{match_name}
\alias{match_name}
\title{Try to Match to State Name}
\usage{
match_name(state)
}
\arguments{
\item{state}{character with state FIPS, Abbreviation, Name, or ANSI}
}
\value{
Name if a match is found or character(0) if no match is found
}
\description{
Searches for an exact match and offers the best match if no exact match
}
\examples{
match_name('NY')
match_name('01')
}
\concept{match}
