% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{mt_county}
\alias{mt_county}
\title{Montana County Data}
\usage{
data('mt_county')
}
\value{
sf tibble with one observation for each county in Montana
}
\description{
\itemize{
\item GEOID: Geographic Identifier
\item NAME: Name of County
\item pop: total population
\item pop_white: total population, Non-Hispanic White
\item pop_black: total population, Non-Hispanic Black
\item pop_hisp: total population, Hispanic
\item pop_aian: total population, Non-Hispanic American Indian and Alaskan Native
\item pop_asian: total population, Non-Hispanic Asian
\item pop_nhpi: total population, Non-Hispanic Native Hawaiian and Pacific Islander
\item pop_other: total population, Non-Hispanic Other
\item pop_two: total population, Non-Hispanic Two Plus Races
\item vap: voting age population
\item vap_white: voting age population, Non-Hispanic White
\item vap_black: voting age population, Non-Hispanic Black
\item vap_hisp: voting age population, Hispanic
\item vap_aian: voting age population, Non-Hispanic American Indian and Alaskan Native
\item vap_asian: voting age population, Non-Hispanic Asian
\item vap_nhpi: voting age population, Non-Hispanic Native Hawaiian and Pacific Islander
\item vap_other: voting age population, Non-Hispanic Other
\item vap_two: voting age population, Non-Hispanic Two Plus Races
\item geometry: sf geometry
}
}
\examples{
data('mt_county')
}
\concept{data}
