% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recode_abb_ansi}
\alias{recode_abb_ansi}
\title{Recode Abb from ANSI}
\usage{
recode_abb_ansi(.data, .ansi)
}
\arguments{
\item{.data}{data.frame or tibble}

\item{.ansi}{column with state ansi}
}
\value{
.data with column .ansi replaced with state abbreviation
}
\description{
Replaces state ansi with state abbreviation
}
\examples{
data('stata')
stata \%>\% recode_ansi_abb(ansi)
}
