% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recode_abb_fips}
\alias{recode_abb_fips}
\title{Recode Abb by FIPS}
\usage{
recode_abb_fips(.data, .fips)
}
\arguments{
\item{.data}{data.frame or tibble}

\item{.fips}{column with state fips}
}
\value{
.data with column .fips replaced with state abb
}
\description{
Replaces state fips with state abb
}
\examples{
data('stata')
stata |> recode_abb_fips(fips)
}
\concept{recode}
