% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recode_abb_name}
\alias{recode_abb_name}
\title{Recode Abb by Name}
\usage{
recode_abb_name(.data, .name)
}
\arguments{
\item{.data}{data.frame or tibble}

\item{.name}{column with state name}
}
\value{
.data with column .name replaced with abbreviation
}
\description{
Replaces state name with state abbreviation
}
\examples{
data('stata')
stata \%>\% recode_abb_name(name)
}
\concept{recode}
