% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recode_ansi_abb}
\alias{recode_ansi_abb}
\title{Recode ANSI as abb}
\usage{
recode_ansi_abb(.data, .ansi)
}
\arguments{
\item{.data}{data.frame or tibble}

\item{.abb}{column with state ansi}
}
\value{
.data with column .ansi replaced with state abbreviation
}
\description{
Replaces state ansi with state abbreviation
}
\examples{
data('stata')
stata \%>\% recode_ansi_abb(ansi)
}
