% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recode_ansi_name}
\alias{recode_ansi_name}
\title{Recode ANSI as Name}
\usage{
recode_ansi_name(.data, .ansi)
}
\arguments{
\item{.data}{data.frame or tibble}

\item{.abb}{column with state ansi}
}
\value{
.data with column .ansi replaced with state name
}
\description{
Replaces state ansi with state name
}
\examples{
data('stata')
stata \%>\% recode_ansi_name(name)
}
