% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recode_fips_abb}
\alias{recode_fips_abb}
\title{Recode FIPS by Abb}
\usage{
recode_fips_abb(.data, .abb)
}
\arguments{
\item{.data}{data.frame or tibble}

\item{.abb}{column with state abbrevaition}
}
\value{
.data with column .abb replaced with state name
}
\description{
Replaces state abbreviation with state fips
}
\examples{
data('stata')
stata |> recode_fips_abb(abb)
}
\concept{recode}
