% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recode_fips_ansi}
\alias{recode_fips_ansi}
\title{Recode FIPS as ANSI}
\usage{
recode_fips_ansi(.data, .fips)
}
\arguments{
\item{.data}{data.frame or tibble}

\item{.fips}{column with state fips}
}
\value{
.data with column .fips replaced with state ansi
}
\description{
Replaces state fips with state ansi
}
\examples{
data('stata')
stata \%>\% recode_fips_abb(fips)
}
