% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recode_fips_name}
\alias{recode_fips_name}
\title{Recode FIPS as Name}
\usage{
recode_fips_name(.data, .fips)
}
\arguments{
\item{.data}{data.frame or tibble}

\item{.fips}{column with state fips}
}
\value{
.data with column .fips replaced with state name
}
\description{
Replaces state fips with state name
}
\examples{
data('stata')
stata \%>\% recode_fips_name(fips)
}
