% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recode_name_abb}
\alias{recode_name_abb}
\title{Recode Name as Abb}
\usage{
recode_name_abb(.data, name)
}
\arguments{
\item{.data}{data.frame or tibble}

\item{.name}{column with state abbreviation}
}
\value{
.data with column .name replaced with abbreviation
}
\description{
Replaces state name with state abbreviation
}
\examples{
data('stata')
stata \%>\% recode_name_abb(name)
}
