% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recode_name_ansi}
\alias{recode_name_ansi}
\title{Recode Name by ANSI}
\usage{
recode_name_ansi(.data, .ansi)
}
\arguments{
\item{.data}{data.frame or tibble}

\item{.ansi}{column with state ansi}
}
\value{
.data with column .ansi replaced with state name
}
\description{
Replaces state ansi with state name
}
\examples{
data('stata')
stata |> recode_name_ansi(name)
}
\concept{recode}
