% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recode_name_fips}
\alias{recode_name_fips}
\title{Recode Name by FIPS}
\usage{
recode_name_fips(.data, .fips)
}
\arguments{
\item{.data}{data.frame or tibble}

\item{.fips}{column with state fips}
}
\value{
.data with column .fips replaced with state name
}
\description{
Replaces state fips with state name
}
\examples{
data('stata')
stata |> recode_name_fips(fips)
}
\concept{recode}
