% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{stata}
\alias{stata}
\title{stata (State Data)}
\usage{
data('stata')
}
\value{
tibble with state identifying information
}
\description{
tibble with columns:
\itemize{
\item fips: Federal Information Processing Standards codes
\item abb: two letter postal abbreviations
\item name: title case state name
\item ansi: American National Standards Institute codes
\item region: Census Regions (for 50 states and D.C.)
\item division: Census Divisions (for 50 states and D.C.)
}
}
\examples{
data('stata')
}
\concept{data}
