% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.R
\name{match_abb}
\alias{match_abb}
\title{Try to Match to State Abbreviation}
\usage{
match_abb(state)
}
\arguments{
\item{state}{character with state FIPS, Abbreviation, Name, or ANSI}
}
\value{
Abbreviation if a match is found or character(0) if no match is found
}
\description{
Searches for an exact match and offers the best match if no exact match
}
\examples{
match_abb('NY')
match_abb('01')
}
\concept{match}
