% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.R
\name{match_fips}
\alias{match_fips}
\title{Try to Match to State FIPS}
\usage{
match_fips(state)
}
\arguments{
\item{state}{character with state FIPS, Abbreviation, Name, or ANSI}
}
\value{
FIPS code if a match is found or character(0) if no match is found
}
\description{
Searches for an exact match and offers the best match if no exact match
}
\examples{
match_fips('NY')
match_fips('01')
}
\concept{match}
