% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recode_abb_ansi}
\alias{recode_abb_ansi}
\title{Recode Abb as ANSI}
\usage{
recode_abb_ansi(.data, .abb)
}
\arguments{
\item{.data}{data.frame or tibble}

\item{.abb}{column with state abbrevaition}
}
\value{
.data with column .abb replaced with state ansi
}
\description{
Replaces state abbreviation with state ansi
}
\examples{
data('stata')
stata \%>\% recode_abb_ansi(abb)
}
