% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recode_ansi_abb}
\alias{recode_ansi_abb}
\title{Recode ANSI by Abb}
\usage{
recode_ansi_abb(.data, .abb)
}
\arguments{
\item{.data}{data.frame or tibble}

\item{.abb}{column with state abbrevaition}
}
\value{
.data with column .abb replaced with state ansi
}
\description{
Replaces state abbreviation with state ansi
}
\examples{
data('stata')
stata |> recode_ansi_abb(abb)
}
\concept{recode}
