% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recode_ansi_fips}
\alias{recode_ansi_fips}
\title{Recode ANSI as FIPS}
\usage{
recode_ansi_fips(.data, .ansi)
}
\arguments{
\item{.data}{data.frame or tibble}

\item{.abb}{column with state ansi}
}
\value{
.data with column .ansi replaced with state fips
}
\description{
Replaces state ansi with state fips
}
\examples{
data('stata')
stata \%>\% recode_ansi_fips(ansi)
}
