% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recode_ansi_name}
\alias{recode_ansi_name}
\title{Recode ANSI by Name}
\usage{
recode_ansi_name(.data, .name)
}
\arguments{
\item{.data}{data.frame or tibble}

\item{.name}{column with state name}
}
\value{
.data with column .name replaced with ansi
}
\description{
Replaces state name with state ansi
}
\examples{
data('stata')
stata \%>\% recode_ansi_name(name)
}
\concept{recode}
