% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recode_fips_ansi}
\alias{recode_fips_ansi}
\title{Recode FIPS from ANSI}
\usage{
recode_fips_ansi(.data, .ansi)
}
\arguments{
\item{.data}{data.frame or tibble}

\item{.ansi}{column with state ansi}
}
\value{
.data with column .ansi replaced with state fips
}
\description{
Replaces state ansi with state fips
}
\examples{
data('stata')
stata \%>\% recode_ansi_fips(ansi)
}
