% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recode_fips_name}
\alias{recode_fips_name}
\title{Recode FIPS by Name}
\usage{
recode_fips_name(.data, .name)
}
\arguments{
\item{.data}{data.frame or tibble}

\item{.name}{column with state name}
}
\value{
.data with column .name replaced with fips
}
\description{
Replaces state name with state fips
}
\examples{
data('stata')
stata \%>\% recode_fips_name(name)
}
\concept{recode}
