% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recode_name_abb}
\alias{recode_name_abb}
\title{Recode Name by Abb}
\usage{
recode_name_abb(.data, .abb)
}
\arguments{
\item{.data}{data.frame or tibble}

\item{.abb}{column with state abbrevaition}
}
\value{
.data with column .abb replaced with state name
}
\description{
Replaces state abbreviation with state name
}
\examples{
data('stata')
stata \%>\% recode_name_abb(abb)
}
\concept{recode}
