% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recode_name_ansi}
\alias{recode_name_ansi}
\title{Recode Name as ANSI}
\usage{
recode_name_ansi(.data, name)
}
\arguments{
\item{.data}{data.frame or tibble}

\item{.name}{column with state ansi}
}
\value{
.data with column .name replaced with ansi
}
\description{
Replaces state name with state ansi
}
\examples{
data('stata')
stata \%>\% recode_name_ansi(name)
}
