% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recode_name_fips}
\alias{recode_name_fips}
\title{Recode Name as FIPS}
\usage{
recode_name_fips(.data, name)
}
\arguments{
\item{.data}{data.frame or tibble}

\item{.name}{column with state fips}
}
\value{
.data with column .name replaced with fips
}
\description{
Replaces state name with state fips
}
\examples{
data('stata')
stata \%>\% recode_name_fips(name)
}
